Autorization checking and @can declarations for user managing

+++ app/Http/Controllers/UserController.php
@@ -4,7 +4,6 @@ namespace App\Http\Controllers;
 
 use App\Http\Requests\CreateUpdateUserRequest;
 use App\Models\User;
-use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Hash;
 
 class UserController extends Controller
@@ -27,9 +26,12 @@ class UserController extends Controller
      * Show the form for creating a new resource.
      *
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function create()
     {
+        $this->authorize('create');
+
         $title = 'User create';
 
         return view('user.create', compact('title'));
@@ -40,9 +42,12 @@ class UserController extends Controller
      *
      * @param  CreateUpdateUserRequest  $request
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function store(CreateUpdateUserRequest $request)
     {
+        $this->authorize('create');
+
         User::create($request->validated());
 
         return redirect(route('user.index'))->with('created', true);
@@ -64,9 +69,12 @@ class UserController extends Controller
      *
      * @param  User  $user
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function edit(User $user)
     {
+        $this->authorize('update', $user);
+
         $title = 'User edit: '.$user->name;
 
         return view('user.edit', compact('title', 'user'));
@@ -78,12 +86,20 @@ class UserController extends Controller
      * @param  CreateUpdateUserRequest  $request
      * @param  User  $user
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function update(CreateUpdateUserRequest $request, User $user)
     {
-        $data = $request->except('password');
+        $this->authorize('update', $user);
+
+        $data = $request->except('password', 'is_admin');
+        // Update password only if itn't null
         if ($request->input('password')) {
             $data['password'] = Hash::make($request->input('password'));
+        }
+        // Only Admin can set another user as Admin
+        if (auth()->user()->is_admin) {
+            $data['is_admin'] = $request->input('is_admin');
         }
 
         $user->update($data);
@@ -96,9 +112,12 @@ class UserController extends Controller
      *
      * @param  int  $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function destroy($id)
     {
+        $this->authorize('delete', $id);
+
         User::destroy([$id]);
 
         return redirect(route('user.index'))->with('deleted', true);
@@ -123,9 +142,12 @@ class UserController extends Controller
      *
      * @param $id
      * @return \Illuminate\Http\Response
+     * @throws \Illuminate\Auth\Access\AuthorizationException
      */
     public function restore($id)
     {
+        $this->authorize('restore', $id);
+
         User::withTrashed()->where('id', $id)->restore();
 
         return redirect(route('user.deleted'))->with('restored', true);
